requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
numpy==2.4.0
protobuf==6.33.2
setuptools==80.9.0
typing_extensions==4.15.0
